# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Veterinary triage tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List active patients (staff)
      description: Returns the list of patients currently in the waiting room (not yet seen). Requires staff authentication.
      responses:
        "200":
          description: Active patient list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Add a new patient to the queue (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientInput"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /patients/{id}:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update an active patient (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientInput"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Mark a patient as departed (staff)
      description: Soft-removes the patient from the active queue when they have departed without being seen. Departed patients are still retained for reporting and appear in the seen / removed history.
      responses:
        "204":
          description: Patient marked as departed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /patients/{id}/start-consult:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    post:
      operationId: startConsult
      tags: [patients]
      summary: Mark a patient as currently in consult (staff)
      description: |
        Flags the patient as actively being seen by a clinician. The patient
        stays in the active list but is excluded from the public waiting-room
        display and from the drag-to-reorder queue. Use `end-consult` to clear
        the flag, or `mark-seen` to complete the consultation.
      responses:
        "200":
          description: Patient is now in consult
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /patients/{id}/end-consult:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    post:
      operationId: endConsult
      tags: [patients]
      summary: Clear the in-consult flag for a patient (staff)
      description: Returns the patient to the regular active queue without marking them as seen.
      responses:
        "200":
          description: Patient returned to the active queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /patients/{id}/mark-seen:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    post:
      operationId: markPatientSeen
      tags: [patients]
      summary: Mark a patient as seen / consultation complete (staff)
      description: Removes the patient from the active queue and stores them in the seen-history for reporting.
      responses:
        "200":
          description: Patient marked as seen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /patients/{id}/restore:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    post:
      operationId: restorePatient
      tags: [patients]
      summary: Restore a previously seen patient back to the active queue (staff)
      responses:
        "200":
          description: Patient restored to active queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /patients/reorder:
    post:
      operationId: reorderPatients
      tags: [patients]
      summary: Reorder the waiting consultation queue (staff)
      description: |
        Atomically rewrites the consultation order for the waiting queue.
        The first ID in the array becomes consultation order 1 (next to be seen),
        the second becomes 2, etc. The submitted IDs must exactly match the set
        of currently *waiting* patients (active and not in consult); patients
        currently marked as in-consult must be excluded. Mismatches are rejected
        with `409 Conflict` so the client can refresh.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderPatientsInput"
      responses:
        "200":
          description: Reordered active patient list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: The active queue has changed since the page was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/seen:
    get:
      operationId: listSeenPatients
      tags: [patients, reports]
      summary: List seen / removed patients (staff)
      description: Returns the historical record of patients that have been seen or removed.
      responses:
        "200":
          description: Seen patient history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /triage/summary:
    get:
      operationId: getTriageSummary
      tags: [patients]
      summary: Aggregated triage stats (staff)
      description: Returns active patient counts grouped by triage class, plus average wait time.
      responses:
        "200":
          description: Triage summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriageSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  parameters:
    PatientId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TriageClass:
      type: string
      enum:
        - red
        - orange
        - yellow
        - green
        - blue
      description: |
        Standard veterinary triage class.
        red = immediate / life-threatening,
        orange = urgent (within 10-15 min),
        yellow = semi-urgent (within 1 hour),
        green = non-urgent,
        blue = routine / wellness
    Species:
      type: string
      description: Species of the patient (e.g. Dog, Cat, Rabbit, Bird, Reptile, Other)
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        age:
          type: string
          description: Free-text age (e.g. "6 months", "2 years")
        species:
          $ref: "#/components/schemas/Species"
        presentingProblem:
          type: string
        triageClass:
          $ref: "#/components/schemas/TriageClass"
        consultationOrder:
          type: integer
          nullable: true
          minimum: 1
          description: Position in consultation order. 1 = next to be seen.
        caseOwner:
          type: string
          description: Veterinarian or staff member responsible for this case
        arrivedAt:
          type: string
          format: date-time
        seenAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [waiting, in_consult, seen, removed]
        removedAt:
          type: string
          format: date-time
          nullable: true
        inConsult:
          type: boolean
          description: True when the patient is currently being seen by a clinician but the consultation is not yet complete.
        consultStartedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - name
        - age
        - species
        - presentingProblem
        - triageClass
        - consultationOrder
        - caseOwner
        - arrivedAt
        - seenAt
        - removedAt
        - status
        - inConsult
        - consultStartedAt
        - notes
    CreatePatientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: string
        species:
          $ref: "#/components/schemas/Species"
        presentingProblem:
          type: string
          minLength: 1
        triageClass:
          $ref: "#/components/schemas/TriageClass"
        consultationOrder:
          type: integer
          nullable: true
          minimum: 1
        caseOwner:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - name
        - age
        - species
        - presentingProblem
        - triageClass
        - caseOwner
    ReorderPatientsInput:
      type: object
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
          description: Active patient IDs in their new consultation order. First ID becomes order 1.
      required:
        - ids
    UpdatePatientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: string
        species:
          $ref: "#/components/schemas/Species"
        presentingProblem:
          type: string
          minLength: 1
        triageClass:
          $ref: "#/components/schemas/TriageClass"
        consultationOrder:
          type: integer
          nullable: true
          minimum: 1
        caseOwner:
          type: string
        notes:
          type: string
          nullable: true
    TriageSummary:
      type: object
      properties:
        totalActive:
          type: integer
          minimum: 0
        averageWaitMinutes:
          type: number
          minimum: 0
        longestWaitMinutes:
          type: number
          minimum: 0
        countsByClass:
          type: object
          properties:
            red:
              type: integer
              minimum: 0
            orange:
              type: integer
              minimum: 0
            yellow:
              type: integer
              minimum: 0
            green:
              type: integer
              minimum: 0
            blue:
              type: integer
              minimum: 0
          required: [red, orange, yellow, green, blue]
        nextPatientName:
          type: string
          nullable: true
          description: First name of the next patient to be seen (consultation order 1)
      required:
        - totalActive
        - averageWaitMinutes
        - longestWaitMinutes
        - countsByClass
        - nextPatientName
...