# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Veterinary triage tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from "./patientStatus";
import type { Species } from "./species";
import type { TriageClass } from "./triageClass";

export interface Patient {
  id: string;
  name: string;
  /** Free-text age (e.g. "6 months", "2 years") */
  age: string;
  species: Species;
  presentingProblem: string;
  triageClass: TriageClass;
  /**
   * Position in consultation order. 1 = next to be seen.
   * @minimum 1
   */
  consultationOrder: number | null;
  /** Veterinarian or staff member responsible for this case */
  caseOwner: string;
  arrivedAt: Date;
  seenAt: Date | null;
  status: PatientStatus;
  removedAt: Date | null;
  /** True when the patient is currently being seen by a clinician but the consultation is not yet complete. */
  inConsult: boolean;
  consultStartedAt: Date | null;
  notes: string | null;
}